Renamed sentinelhub.download.request module into sentinelhub.download.models
