+++ sentinelhub/download/__init__.py
@@ -3,7 +3,7 @@ A download part of the package
 """
 
 from .client import DownloadClient
-from .request import DownloadRequest
+from .models import DownloadRequest
 from .sentinelhub_client import SentinelHubDownloadClient
 from .sentinelhub_statistical_client import SentinelHubStatisticalDownloadClient
 from .session import SentinelHubSession, SessionSharing, SessionSharingThread, collect_shared_session
