+++ sentinelhub/download/models.py
@@ -1,5 +1,5 @@
 """
-Module implementing DownloadRequest class
+Module implementing model classes to store download-related parameters and data.
 """
 from __future__ import annotations
