+++ sentinelhub/download/session.py
@@ -18,7 +18,7 @@ from requests_oauthlib import OAuth2Session
 
 from ..config import SHConfig
 from ..download.handlers import fail_user_errors, retry_temporary_errors
-from ..download.request import DownloadRequest
+from ..download.models import DownloadRequest
 from ..exceptions import SHUserWarning
 from ..type_utils import JsonDict
